Fix for error when calling public method getCallsRemaining()
getCallsRemaining() fails with PHP Warning `explode() expects parameter 2 to be string, array given`. This is due to `$this->last_response_headers['X-Shopify-Shop-Api-Call-Limit']` on getCallLimitHeaderValue() returning array an not string.
PR is to select the first element of that array, which the string we're looking for: "<callsMade>/<callsLimit>". That string format will feed the explode() function properly.